Added "double decode"
The "double decode" activity tests decoding two video streams to
two TextureViews.  The video continues to play when the screen is
rotated, avoiding the (potentially expensive) decoder teardown
and re-init.

It turns out there's a bug that prevents the TextureViews from
updating after the rotation, so this doesn't work quite right.