Reshuffle thread start in two activities
RecordFBOActivity and HardwareScalerActivity were starting the
render thread in surfaceCreated(), and stopping it in onPause().
The asymmetry felt weird, so we now stop the thread in
surfaceDestroyed().

The surfaceDestroyed() callback wasn't firing because we were
releasing the Surface in WindowSurface.  This also felt weird,
because the framework should be allowed to manage the life span
of a SurfaceView's Surface, so we no longer release it.

Should be no net change in behavior.