Handle content-generation failures better
If we somehow got a software codec -- maybe we're on an emulator --
the creation of the generated videos would fail.  This killed the
app shortly after it started, leaving an obscure failure in the
log file.

We now trap the failure and put up an exception dialog warning
that not all features may be available.  If it looks like we're
trying to use the software AVC codec, we make that explicit in
the error dialog.