Fix video player crash
Sometimes the video player will crash if you rotate the device
while a video is playing, because we aren't waiting for video
playback to stop before returning from onPause().  This causes the
TextureView to get torn down, which makes the MediaCodec unhappy
the next time it tries to decode a frame.

Now we wait for the playback thread to stop.