This should fix the freezing cameraPreview if the app is put to sleep with the power button and is woken up later. The static sSurfaceHolder stores the surfaceHolder and onResume() checks if it is still set. There are two cases: 1. If the app is put into background the surfaceHolder gets destroyed correctly (see surfaceDestroyed()) and the static variable is set to null. 2. If the power button is pressed surfaceDestroyed() callback is never called because the surface is not destroyed.
In case number 1 we do not have to do anything because our app just goes trough the inital setup with the help of the Surface Callback
In case number 2 we still have the surface but we released() everything else, therefore we have to setup() all the other things after opening the camera (EGL, mCameraTexture, mEncoderSurface, mCircEncoder...)

This works for me and I hope it will work for others as well.

I could not really test it because I cannot build this directory but I used this exact solution in my own app which is built on grafika/Continuous Capture for your changes. Lines starting